requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
flask-restx==1.3.0
Werkzeug==2.3.7

# Production WSGI Server
gunicorn>=20.1.0

# Computer Vision & Image Processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Data Science & Machine Learning
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Face Recognition (if needed - can be heavy)
# dlib==19.24.2

# Data Visualization
matplotlib>=3.7.0

# HTTP Requests
requests>=2.31.0

# Environment Management
python-dotenv>=1.0.0

# Database
supabase>=2.15.0

# Additional dependencies for production
psutil>=5.9.0 